+++ R/colour.R
@@ -1,8 +1,8 @@
 
 
-
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#' Hash a string
+#' Hash a string. This hash is used for colour lookup.
+#' 
 #' @param str single string
 #' @return integer
 #' @noRd
@@ -12,71 +12,37 @@ djb2_hash <- function(str) {
 }
 
 
-
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-# Generate a big switch/case statement in C code for converting
-# an R colour to an integer.
-# Rather than doing string compares, this is comparing hashes of the strings.
-# Not really sure that an integer hash switch/case is any faster/slower
-# than a strncmp() for this use case.
-#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-if (FALSE) {
-  cols <- colours()
-  zz <- purrr::map_int(cols, ~sum(utf8ToInt(.x)))
-
-  normalize_colour <- function(colour) {
-    res <- t(col2rgb(colour, alpha = TRUE))
-    rgb(res[,1], res[, 2], res[,3], res[,4], maxColorValue = 255)
-  }
-
-
-
-  sink("col.txt")
-  for (colour in colours()) {
-    cat('  case ', djb2_hash(colour), ':\n    res = ', colour_to_integer(normalize_col(colour)), ';\n    break;\n', sep="")
-  }
-  sink()
-}
-
-
-#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#' Convert a colour as a character string  into an integer value representing RGBA
+#' Convert colours (R colours and hex colours) into packed colours 
+#' (integer values containing RGBA bytes)
 #'
-#' @param colour character vector of R colour names (including hex colours)
+#' @param colours character vector of R colour names and hex colours e.g. 
+#'        \code{c('red', 'white', NA, 'transparent', '#12345678')}
 #'
-#' @return Integer value representing RGBA colour
-#' @importFrom grDevices col2rgb
+#' @return Integer vector. Each integer value contains RGBA bytes.
 #' @export
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-colour_to_integer <- function(colour) {
-  .Call(col_to_int_, colour)
+str_cols_to_packed_cols <- function(colours) {
+  .Call(colours_to_packed_cols_, colours)
 }
 
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#' Convert integer values to hex colours
+#' Convert packed colours (integer values containing RGBA bytes) to hex colours
 #'
-#' @param ints integer values
+#' @param packed_cols integer values each containing packed RGBA colour information
 #' @return character vector of hex colours
 #'
 #' @export
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-integer_to_colour <- function(ints) {
-  .Call(int_to_col_, ints)
+packed_cols_to_hex_cols <- function(packed_cols) {
+  .Call(packed_cols_to_hexcolours_, packed_cols)
 }
 
 
-if (FALSE) {
-  
-  colour_to_integer("#FFFFFF00") |> integer_to_colour()
-  colour_to_integer("transparent") |> integer_to_colour()
-  colour_to_integer(NA_character_) |> integer_to_colour()
-  
-  col2rgb(c("transparent", "#FFFFFF00", NA_character_, "white"), alpha = T)
-  colour_to_integer(c("transparent", "#FFFFFF00", NA_character_, "white"))
-  
-}
-
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+# Inline benchmarks
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 if (FALSE) {
   
   nr <- nr_new(500, 500)
@@ -85,28 +51,24 @@ if (FALSE) {
 
   bench::mark(  
     farver::encode_native(colours()),
-    colour_to_integer(colours())
+    str_cols_to_packed_cols(colours())
   )
   
   bench::mark(  
     farver::encode_native(colour_hex),
-    colour_to_integer(colour_hex)
+    str_cols_to_packed_cols(colour_hex)
   )
   
   
   res <- farver::decode_native(colour_ints)
   res <- ifelse(nchar(res) == 7, paste0(res, "FF"), res)
-  identical(res, int_to_col(colour_ints))
+  identical(res, packed_cols_to_hex_cols(colour_ints))
   
   bench::mark(  
     farver::decode_native(colour_ints),
-    integer_to_colour(colour_ints),
-    int_to_col(colour_ints),
+    packed_cols_to_hex_cols(colour_ints),
     check = FALSE
   )
-  
-  
-  
 }