+++ src/nr-utils.h
@@ -2,5 +2,5 @@
 
 
 void assert_nativeraster(SEXP nr_) ;
-int *as_int32_vec(SEXP vec_, int N, int *do_free);
+int *as_int32_vec(SEXP vec_, int N, bool *do_free);
 int calc_max_length(int count, ...);